requirements.txt
streamlit
matplotlib
pillow
reportlab